+++ src/mercator/mercator/static/js/Packages/Mercator/2016/Proposal/Proposal.ts
@@ -1,0 +1,21 @@
+/// <reference path="../../../../../lib/DefinitelyTyped/angularjs/angular.d.ts"/>
+/// <reference path="../../../../../lib/DefinitelyTyped/moment/moment.d.ts"/>
+
+import * as AdhConfig from "../../../Config/Config";
+
+var pkgLocation = "/Mercator/2016/Proposal";
+
+
+export var createDirective = (
+    adhConfig : AdhConfig.IService
+) => {
+    return {
+        restrict: "E",
+        templateUrl: adhConfig.pkg_path + pkgLocation + "/Create.html",
+        scope: {},
+        link: (scope) => {
+            console.log(scope);
+
+        }
+    };
+};
