--- src/mercator/mercator/static/js/Packages/MercatorProposal2016/MercatorProposal2016.ts
@@ -1,20 +1,0 @@
-/// <reference path="../../../lib/DefinitelyTyped/angularjs/angular.d.ts"/>
-/// <reference path="../../../lib/DefinitelyTyped/moment/moment.d.ts"/>
-
-import * as AdhConfig from "../Config/Config";
-
-var pkgLocation = "/MercatorProposal2016";
-
-export var createDirective = (
-    adhConfig : AdhConfig.IService
-) => {
-    return {
-        restrict: "E",
-        templateUrl: adhConfig.pkg_path + pkgLocation + "/create.html",
-        scope: {},
-        link: (scope) => {
-            console.log(scope);
-
-        }
-    };
-};
